requirements.txt
alembic==1.4.3
attrs==21.2.0
backcall==0.2.0
bcrypt==4.0.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
cryptography==37.0.4
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==4.4.2
eventkit==0.8.6
falcon==2.0.0
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
ib-insync==0.9.64
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.0
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.4
nest-asyncio==1.4.3
numpy==1.19.3
packaging==20.9
pandas==1.1.4
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.10.0
py-common-utils-gh @ git+https://github.com/clockworcarry/py_common_utils.git@cf5b1e5a7cac0b134928aaa773760a10366b0118
pybind11==2.6.2
pycparser==2.21
pydantic==1.8
Pygments==2.7.3
PyJWT==2.4.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
pytz==2020.1
pyzmq==20.0.0
requests==2.24.0
rope==0.18.0
semantic-version==2.10.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.25.11
uvicorn==0.13.4
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
